Use browserify for data-explorer. Fixes #917.
