+++ data_explorer/templates/base.html
@@ -44,7 +44,7 @@
     </script>
 
     {# For the list of files included in this built script, see gulpfile.js #}
-    <script src="{% static 'frontend/built/js/common/base.min.js' %}"></script>
+    <script src="{% static 'frontend/built/js/common/vendor.min.js' %}"></script>
     <![endif]>
     <link href='https://fonts.googleapis.com/css?family=Lato:300,400,700,400italic,700italic' rel='stylesheet' type='text/css'>
     {% block head %}
