+++ data_explorer/templates/index.html
@@ -520,6 +520,7 @@
 
 <![if gt IE 8]>
 {# For the list of files included in this built script, see gulpfile.js #}
+<script src= "{% static 'frontend/built/js/data-explorer/vendor.min.js' %}"></script>
 <script src= "{% static 'frontend/built/js/data-explorer/index.min.js' %}"></script>
 <![endif]>
